Extract interface from AssetManager
